+++ tests/Unit/Search/ItemSearchProviderTest.php
@@ -2,37 +2,36 @@
 
 namespace OCA\News\Search;
 
-use OCA\News\Db\Folder;
-use OCA\News\Service\FeedServiceV2;
-use OCA\News\Service\FolderServiceV2;
-use OCA\News\Service\OpmlService;
-use OCA\News\Utility\OPMLExporter;
+use OCA\News\Db\Item;
+use OCA\News\Db\ListType;
+use OCA\News\Service\ItemServiceV2;
 use OCP\IL10N;
 use OCP\IURLGenerator;
 use OCP\IUser;
 use OCP\Search\ISearchQuery;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class FolderSearchProviderTest extends TestCase
+class ItemSearchProviderTest extends TestCase
 {
 
     /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|FolderServiceV2
+     * @var MockObject|ItemServiceV2
      */
-    private $folderService;
+    private $itemService;
 
     /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|IL10N
+     * @var MockObject|IL10N
      */
     private $l10n;
 
     /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|IURLGenerator
+     * @var MockObject|IURLGenerator
      */
     private $generator;
 
     /**
-     * @var FolderSearchProvider
+     * @var ItemSearchProvider
      */
     private $class;
 
@@ -44,40 +43,40 @@ class FolderSearchProviderTest extends TestCase
         $this->generator = $this->getMockBuilder(IURLGenerator::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->folderService = $this->getMockBuilder(FolderServiceV2::class)
+        $this->itemService = $this->getMockBuilder(ItemServiceV2::class)
             ->disableOriginalConstructor()
             ->getMock();
 
-        $this->class = new FolderSearchProvider(
+        $this->class = new ItemSearchProvider(
             $this->l10n,
             $this->generator,
-            $this->folderService
+            $this->itemService
         );
     }
 
     public function testGetId()
     {
-        $this->assertSame('news_folder', $this->class->getId());
+        $this->assertSame('news_item', $this->class->getId());
     }
 
     public function testGetName()
     {
         $this->l10n->expects($this->once())
                    ->method('t')
-                   ->with('News folders')
+                   ->with('News articles')
                    ->willReturnArgument(0);
 
-        $this->assertSame('News folders', $this->class->getName());
+        $this->assertSame('News articles', $this->class->getName());
     }
 
     public function testGetOrderExternal()
     {
-        $this->assertSame(55, $this->class->getOrder('contacts.Page.index', []));
+        $this->assertSame(65, $this->class->getOrder('contacts.Page.index', []));
     }
 
     public function testGetOrderInternal()
     {
-        $this->assertSame(0, $this->class->getOrder('news.page.index', []));
+        $this->assertSame(1, $this->class->getOrder('news.page.index', []));
     }
 
     public function testSearch()
@@ -87,23 +86,39 @@ class FolderSearchProviderTest extends TestCase
         $query = $this->getMockBuilder(ISearchQuery::class)
                      ->getMock();
 
+        $query->expects($this->once())
+                ->method('getCursor')
+                ->willReturn(null);
+        
+        $query->expects($this->once())
+                ->method('getLimit')
+                ->willReturn(10);
+        
         $user->expects($this->once())
-              ->method('getUID')
-              ->willReturn('user');
+                ->method('getUID')
+                ->willReturn('user');
 
         $query->expects($this->once())
               ->method('getTerm')
-              ->willReturn('Term');
+              ->willReturn('some text');
+        
 
-        $folders = [
-            Folder::fromRow(['id' => 1,'name' => 'some_tErm']),
-            Folder::fromRow(['id' => 2,'name' => 'nothing'])
+        $items = [
+            Item::fromRow(['id' => 1,'title' => 'some_tErm', 'body' => 'some text', 'feedId' => 1]),
+            Item::fromRow(['id' => 2,'title' => 'nothing', 'body' => 'some text', 'feedId' => 1])
         ];
 
-        $this->folderService->expects($this->once())
-                            ->method('findAllForUser')
-                            ->with('user')
-                            ->willReturn($folders);
+        $this->itemService->expects($this->once())
+                            ->method('findAllWithFilters')
+                            ->with(
+                                'user',
+                                ListType::ALL_ITEMS,
+                                10,
+                                0,
+                                false,
+                                ['some text'])
+                            ->willReturn($items);
+
 
         $this->l10n->expects($this->once())
             ->method('t')
@@ -112,10 +127,10 @@ class FolderSearchProviderTest extends TestCase
 
         $this->generator->expects($this->once())
                         ->method('imagePath')
-                        ->with('core', 'filetypes/folder.svg')
+                        ->with('core', 'filetypes/text.svg')
                         ->willReturn('folderpath.svg');
 
-        $this->generator->expects($this->once())
+        $this->generator->expects($this->exactly(2))
                         ->method('linkToRoute')
                         ->with('news.page.index')
                         ->willReturn('/news');
@@ -126,7 +141,7 @@ class FolderSearchProviderTest extends TestCase
         $this->assertSame('News', $result['name']);
         $this->assertSame('some_tErm', $entry['title']);
         $this->assertSame('folderpath.svg', $entry['thumbnailUrl']);
-        $this->assertSame('', $entry['subline']);
-        $this->assertSame('/news#/items/folders/1', $entry['resourceUrl']);
+        $this->assertSame('some text', $entry['subline']);
+        $this->assertSame('/news#/items/feeds/1', $entry['resourceUrl']);
     }
 }